+++ src/main/kotlin/lotto/Lotto.kt
@@ -3,7 +3,12 @@ package lotto
 class Lotto(private val numbers: List<Int>) {
     init {
         require(numbers.size == 6) { "[ERROR] Lotto must contain exactly 6 numbers." }
+        require(numbers.toSet().size == numbers.size) { "[ERROR] Lotto numbers must be unique." }
+        require(numbers.all { it in 1..45 }) { "[ERROR] All numbers must be between 1 and 45." }
     }
 
-    // TODO: Implement additional functions
+    // Method to get the numbers
+    fun getNumbers(): List<Int> {
+        return numbers
+    }
 }